Move live Debugger related items to its own debugmanager
